Refactoring on the Screenshots Classes
