+++ Amigula.Domain/Interfaces/IGamesRepository.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Amigula.Domain.Classes;
+﻿using System.Collections.Generic;
 using Amigula.Domain.DTO;
 
 namespace Amigula.Domain.Interfaces
@@ -11,9 +6,5 @@ namespace Amigula.Domain.Interfaces
     public interface IGamesRepository
     {
         IEnumerable<GamesDto> GetGamesList();
-
-        bool FilenameExists(string filenameFullPath);
-        OperationResult CopyFileInPlace(string filename, string destination);
-        OperationResult RenameFile(string oldFilename, string newFilename);
     }
 }
